# api/openapi.yaml
openapi: 3.1.2
info:
  title: Descendence API
  description: Basic api provided by the descendence api server
  version: v0.0.1
servers:
- url: http://127.0.0.1:8080
  description: Local development server for testing
paths:
...
  /api/v1/repos/{id}:
    get:
      operationId: getRepo
      security:
        - bearerAuth: []
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: The repository
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Repo"
        "401":
          description: Missing, malformed, unknown, expired or revoked token
          content:
            application/problem+json:
              schema:
                $ref: "#/components/schemas/Problem"
        "404":
          description: No repository with this id
          content:
            application/problem+json:
              schema:
                $ref: "#/components/schemas/Problem"
...
  /api/v1/jobs:
    get:
      operationId: listJobs
      security:
        - bearerAuth: []
      parameters:
        - name: cursor
          in: query
          schema:
            type: string
        - name: limit
          in: query
          schema:
            type: integer
            minimum: 1
            maximum: 200
            default: 50
      responses:
        "200":
          description: >
            Live jobs ordered by name, keyset-paginated. Ordered by name rather
            than by time because a job list is a catalogue, not a timeline -
            the opposite of the runs list.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/JobList"
        "400":
          description: Malformed cursor, or limit outside 1-200
          content:
            application/problem+json:
              schema:
                $ref: "#/components/schemas/Problem"
        "401":
          description: Missing, malformed, unknown, expired or revoked token
          content:
            application/problem+json:
              schema:
                $ref: "#/components/schemas/Problem"
  /api/v1/jobs/{id}:
    get:
      operationId: getJob
      description: >
        Returns the job, including one whose manifest has been removed - a run
        pointing at a deleted job must still be explainable.
      security:
        - bearerAuth: []
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: The job
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Job"
        "401":
          description: Missing, malformed, unknown, expired or revoked token
          content:
            application/problem+json:
              schema:
                $ref: "#/components/schemas/Problem"
        "404":
          description: No job with this id
          content:
            application/problem+json:
              schema:
                $ref: "#/components/schemas/Problem"
    patch:
      operationId: patchJob
      description: >
        Enable or disable a job.


        `enabled` is the only field of a job the API can change, because it is
        the only one this installation owns rather than the repository. Every
        other field comes from the manifest; to change one, commit the manifest
        (POST /api/v1/repos/{id}/files) and sync.
      security:
        - bearerAuth: []
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/JobPatch"
      responses:
        "200":
          description: The updated job
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Job"
        "400":
          description: Malformed body, or a field other than `enabled` was supplied
          content:
            application/problem+json:
              schema:
                $ref: "#/components/schemas/Problem"
        "401":
          description: Missing, malformed, unknown, expired or revoked token
          content:
            application/problem+json:
              schema:
                $ref: "#/components/schemas/Problem"
        "404":
          description: No live job with this id
          content:
            application/problem+json:
              schema:
                $ref: "#/components/schemas/Problem"
  /api/v1/jobs/{id}/runs:
    post:
      operationId: createJobRun
      description: >
        Trigger a job.


        The request carries no execution detail: what to run is the manifest's
        business. The server resolves the repository's current commit, reads
        the manifest *at that commit*, and records the resolved image, argv and
        commit SHA onto the run - so the branch moving a moment later cannot
        change what this run executes.


        202 with a Location header, like every other way of creating a run.
      security:
        - bearerAuth: []
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
        - $ref: "#/components/parameters/IdempotencyKey"
      responses:
        "202":
          description: >
            Run accepted and queued, with jobId and commitSha already set. A
            repeated Idempotency-Key returns the original run.
          headers:
            Location:
              description: "/api/v1/runs/{id} of the created (or original, if replayed) run"
              schema:
                type: string
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Run"
        "401":
          description: Missing, malformed, unknown, expired or revoked token
          content:
            application/problem+json:
              schema:
                $ref: "#/components/schemas/Problem"
        "404":
          description: No job with this id
          content:
            application/problem+json:
              schema:
                $ref: "#/components/schemas/Problem"
        "409":
          description: >
            The job cannot be run as it stands: it is disabled, its manifest
            has been removed from the repository, the repository has no
            commits, the manifest is missing or invalid at the current
            commit, or the manifest names a runtime that is not defined, not
            yet built, or whose image has been pruned.
          content:
            application/problem+json:
              schema:
                $ref: "#/components/schemas/Problem"
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
  parameters:
    IdempotencyKey:
      name: Idempotency-Key
      in: header
      description: >
        Scoped per principal. A retried request with the same key returns the
        original run instead of creating a second one.
      schema:
        type: string
  schemas:
...
    Problem:
      type: object
      description: RFC 9457 problem details
      properties:
        type:
          type: string
        title:
          type: string
        status:
          type: integer
        detail:
          type: string
      required:
        - type
        - title
        - status
...
    Run:
      type: object
      properties:
        id:
          type: integer
        state:
          type: string
          enum: [queued, running, succeeded, failed, cancelled, lost]
        imageRef:
          type: string
        argv:
          type: array
          items:
            type: string
        timeoutSeconds:
          type: integer
        containerId:
          type: [string, "null"]
        exitCode:
          type: [integer, "null"]
        failureReason:
          type: [string, "null"]
        cancelRequestedAt:
          description: >
            When cancellation was requested. Set before the run is actually
            cancelled - a running run's container still has to be stopped by
            the supervisor - so a run can be `running` with this set, meaning
            "on its way out". Also kept on a run that finished on its own
            before the request landed, which is the case worth being able to
            explain afterwards.
          type: [string, "null"]
          format: date-time
        queuedAt:
          type: string
          format: date-time
        startedAt:
          type: [string, "null"]
          format: date-time
        finishedAt:
          type: [string, "null"]
          format: date-time
        jobId:
          description: >
            The job this run came from, or null for an ad-hoc run created by
            POST /api/v1/runs. A job may since have been removed from its
            repository; the run still points at it, so that what ran remains
            explainable.
          type: [integer, "null"]
        commitSha:
          description: >
            The exact commit the job's manifest and script were read from,
            resolved when the run was created and never re-resolved. Together
            with jobId this is what makes a past run reproducible: check this
            commit out and you have precisely what executed. Null for an
            ad-hoc run.
          type: [string, "null"]
        runtimeId:
          description: >
            Set when the job named a runtime rather than an image directly
            (task 4.6). Resolved once, at creation, from the runtime's
            current image digest - rebuilding the runtime afterwards never
            changes this run's imageDigest.
          type: [integer, "null"]
        imageDigest:
          description: The runtime's image digest at the moment this run was created. Null for an ad-hoc run or a job that names an image directly.
          type: [string, "null"]
      required:
        - id
        - state
        - imageRef
        - argv
        - timeoutSeconds
        - queuedAt
...
    Repo:
      type: object
      description: >
        A bare git repository holding job definitions. The api is its sole
        writer; the supervisor only reads a script blob at a commit a run has
        pinned.
      properties:
        id:
          type: integer
        name:
          description: >
            Unique, and used as the directory name under GIT_REPO_DIR, so it
            is constrained to letters, digits, dots, dashes and underscores.
          type: string
        path:
          description: Where the repository lives on disk.
          type: string
        kind:
          type: string
          enum: [local, external]
        defaultBranch:
          description: >
            The branch a scan reads and a job run resolves against. Recorded
            per repository rather than assumed globally.
          type: string
        remoteUrl:
          description: Always null for a local repository. External repositories are deferred.
          type: [string, "null"]
        lastSyncedAt:
          description: >
            What the last scan saw. Reported for the operator's benefit and
            never used as an input to the next scan, which always re-reads
            HEAD - trusting a stored "last seen" would make a half-failed sync
            invisible to the one after it.
          type: [string, "null"]
          format: date-time
        lastSyncedCommitSha:
          type: [string, "null"]
      required:
        - id
        - name
        - path
        - kind
        - defaultBranch
...
    Job:
      type: object
      description: >
        A job is a script's interface - what to run, how to invoke it, and
        (from later phases) what parameters it takes and how they are
        presented. It is authored as a `<name>.job.yaml` manifest in git,
        versioned alongside the script it describes.


        This resource is a **projection** of that manifest, rebuilt by a scan.
        Every field below except `enabled` is owned by git and rewritten by the
        next sync, which is why there is no endpoint that edits one: to change
        a job, commit its manifest.
      properties:
        id:
          type: integer
        repoId:
          type: integer
        name:
          description: Unique among live jobs, and how a job is addressed on the command line.
          type: string
        description:
          type: [string, "null"]
        manifestPath:
          description: >
            Where the manifest lives in the repository. Immutable for a given
            job: a job is keyed on (repository, manifest path), so moving a
            manifest soft-deletes one job and creates another.
          type: string
        scriptPath:
          description: >
            Repository-root-relative, resolved from the manifest's own
            directory - a manifest and its script are a unit that can be moved
            or copied together.
          type: string
        command:
          description: >
            Null in the usual case, meaning argv is the script's own path and
            its shebang chooses the interpreter. Set only when the manifest
            overrides the invocation.
          type: [array, "null"]
          items:
            type: string
        imageRef:
          description: >
            A plain OCI reference, used as-is. Nullable: a manifest names
            exactly one of `image:` or `runtime:` (task 4.6), so exactly one
            of imageRef/runtimeId is set.
          type: [string, "null"]
        runtimeId:
          description: >
            Set when the manifest names a runtime with `runtime:` instead of
            an image directly. A run created from this job resolves the
            runtime's current image digest at creation time and pins it -
            rebuilding the runtime afterwards never changes what an
            already-created run executes.
          type: [integer, "null"]
        timeoutSeconds:
          description: Null when the manifest does not say, in which case the platform default applies.
          type: [integer, "null"]
        enabled:
          description: >
            The one field git does not own. A sync never changes it, so
            pausing a misbehaving job is not something the next scan undoes.
          type: boolean
        syncedCommitSha:
          description: Which commit this row was built from.
          type: string
        deletedAt:
          description: >
            Set when the manifest has been removed from the repository. Such a
            job cannot be run and does not appear in listings, but the record
            survives so that runs which used it remain explainable. Restoring
            the manifest resurrects this same job, with its run history.
          type: [string, "null"]
          format: date-time
      required:
        - id
        - repoId
        - name
        - manifestPath
        - scriptPath
        - enabled
        - syncedCommitSha
    JobList:
      type: object
      properties:
        items:
          description: Live jobs only, ordered by name.
          type: array
          items:
            $ref: "#/components/schemas/Job"
        nextCursor:
          description: Opaque cursor for the next page. Absent (or null) on the last page.
          type: [string, "null"]
      required:
        - items
    JobPatch:
      type: object
      description: >
        The only mutation the API performs on a job. Everything else about a
        job is defined by its manifest in git.
      properties:
        enabled:
          type: boolean
      required:
        - enabled
...