# internal/store/models.go
// Code generated by sqlc. DO NOT EDIT.
// versions:
//   sqlc v1.31.1

package store

import (
	"github.com/jackc/pgx/v5/pgtype"
)

type Audit struct {
	ID          int64              `json:"id"`
	PrincipalID pgtype.Int8        `json:"principal_id"`
	Action      string             `json:"action"`
	Target      pgtype.Text        `json:"target"`
	Ts          pgtype.Timestamptz `json:"ts"`
	DetailJson  []byte             `json:"detail_json"`
}

type Job struct {
	ID              int64              `json:"id"`
	RepoID          int64              `json:"repo_id"`
	RuntimeID       pgtype.Int8        `json:"runtime_id"`
	ManifestPath    string             `json:"manifest_path"`
	Name            string             `json:"name"`
	Enabled         bool               `json:"enabled"`
	CreatedAt       pgtype.Timestamptz `json:"created_at"`
	Description     pgtype.Text        `json:"description"`
	ScriptPath      string             `json:"script_path"`
	Command         []string           `json:"command"`
	ImageRef        pgtype.Text        `json:"image_ref"`
	TimeoutSeconds  pgtype.Int4        `json:"timeout_seconds"`
	SyncedCommitSha string             `json:"synced_commit_sha"`
	SyncedAt        pgtype.Timestamptz `json:"synced_at"`
	DeletedAt       pgtype.Timestamptz `json:"deleted_at"`
}

type Principal struct {
	ID        int64              `json:"id"`
	Kind      string             `json:"kind"`
	Name      string             `json:"name"`
	TokenHash []byte             `json:"token_hash"`
	TokenHint pgtype.Text        `json:"token_hint"`
	Scopes    []string           `json:"scopes"`
	CreatedAt pgtype.Timestamptz `json:"created_at"`
	ExpiresAt pgtype.Timestamptz `json:"expires_at"`
	RevokedAt pgtype.Timestamptz `json:"revoked_at"`
}

type Repo struct {
	ID                  int64              `json:"id"`
	Name                string             `json:"name"`
	Path                string             `json:"path"`
	Kind                string             `json:"kind"`
	RemoteUrl           pgtype.Text        `json:"remote_url"`
	CreatedAt           pgtype.Timestamptz `json:"created_at"`
	DefaultBranch       string             `json:"default_branch"`
	LastSyncedAt        pgtype.Timestamptz `json:"last_synced_at"`
	LastSyncedCommitSha pgtype.Text        `json:"last_synced_commit_sha"`
}

type Run struct {
	ID                int64              `json:"id"`
	PrincipalID       int64              `json:"principal_id"`
	State             string             `json:"state"`
	IdempotencyKey    pgtype.Text        `json:"idempotency_key"`
	ImageRef          string             `json:"image_ref"`
	Argv              []string           `json:"argv"`
	TimeoutSeconds    int32              `json:"timeout_seconds"`
	ContainerID       pgtype.Text        `json:"container_id"`
	ExitCode          pgtype.Int4        `json:"exit_code"`
	FailureReason     pgtype.Text        `json:"failure_reason"`
	CancelRequestedAt pgtype.Timestamptz `json:"cancel_requested_at"`
	QueuedAt          pgtype.Timestamptz `json:"queued_at"`
	StartedAt         pgtype.Timestamptz `json:"started_at"`
	FinishedAt        pgtype.Timestamptz `json:"finished_at"`
	JobID             pgtype.Int8        `json:"job_id"`
	CommitSha         pgtype.Text        `json:"commit_sha"`
	RuntimeID         pgtype.Int8        `json:"runtime_id"`
	ImageDigest       pgtype.Text        `json:"image_digest"`
	ParamsJson        []byte             `json:"params_json"`
	LogsPrunedAt      pgtype.Timestamptz `json:"logs_pruned_at"`
}

type RunLog struct {
	RunID      int64              `json:"run_id"`
	Seq        int64              `json:"seq"`
	Stream     string             `json:"stream"`
	Ts         pgtype.Timestamptz `json:"ts"`
	ByteOffset int64              `json:"byte_offset"`
	ByteLength int32              `json:"byte_length"`
}

type Runtime struct {
	ID            int64              `json:"id"`
	Name          string             `json:"name"`
	BaseImage     string             `json:"base_image"`
	SysPackages   []string           `json:"sys_packages"`
	LangManifest  pgtype.Text        `json:"lang_manifest"`
	ImageDigest   pgtype.Text        `json:"image_digest"`
	BuildStatus   string             `json:"build_status"`
	BuiltAt       pgtype.Timestamptz `json:"built_at"`
	CreatedAt     pgtype.Timestamptz `json:"created_at"`
	Lang          string             `json:"lang"`
	InputHash     string             `json:"input_hash"`
	BuildError    pgtype.Text        `json:"build_error"`
	ImagePrunedAt pgtype.Timestamptz `json:"image_pruned_at"`
}

type Schedule struct {
	ID        int64              `json:"id"`
	JobID     int64              `json:"job_id"`
	CronExpr  string             `json:"cron_expr"`
	Timezone  string             `json:"timezone"`
	NextDueAt pgtype.Timestamptz `json:"next_due_at"`
	Enabled   bool               `json:"enabled"`
	CreatedAt pgtype.Timestamptz `json:"created_at"`
}